+++ src/turtletorque3/turtletorque3_motor_driver.cpp
@@ -160,7 +160,7 @@ bool TurtleTorque3MotorDriver::readEncoder(int32_t &left_value, int32_t &right_v
   return true;
 }
 
-bool TurtleTorque3MotorDriver::readCurrent(int32_t &left_value, int32_t &right_value)
+bool TurtleTorque3MotorDriver::readCurrent(int16_t &left_value, int16_t &right_value)
 {
   int dxl_comm_result = COMM_TX_FAIL;              // Communication result
   bool dxl_addparam_result = false;                // addParam result
